move auth user actions to /api/auth/user/...
